Paths should be a bit more Enumerable-y
The use of the #points array and the need to freeze and dup it on init
is beginning to nag at me.

There's also the realisation that Arc as it stands is also pathlike -
it's a start point plus (currently) a set of cubic beziers. I'm
wondering how to unify the relevant bits of interface there.