Add pointlike Curve to represent arbitrary curves
The curve portion of an Arc is conceptually one point-like thing (as
implemented in e.g. SVG), but may have to be represented as multiple
cubic Bezier curves (e.g. in PDF).

Curve is pointlike (representing the end point in the same way as a
CubicBezier), but can also return the one-or-more CubicBezier representation of
itself.